How to use DocumenterCitations.jl

DocumenterCitations.jl uses Bibliography.jl to add support for BibTeX citations and references in documentation pages generated by Documenter.jl.

DocumenterCitations.jl is still in early development so please open issues if you encounter any bugs or pain points, would like to see a new feature, or if you have any questions.

Installation instructions

You can install the latest version of DocumenterCitations.jl using the built-in package manager (accessed by pressing ] in the Julia command prompt) to add the package and instantiate/build all dependencies

julia>]
(v1.5) pkg> add DocumenterCitations

Telling Documenter.jl about your bibliography

You just need a BibTeX .bib file to get started! Let's call it some_references.bib. You want to create a CitationBibliography and pass it to makedocs:

using DocumenterCitations

bib = CitationBibliography("some_references.bib")
makedocs(bib, ...)

How to get a sorted References list

In order to get a sorted references list you can pass a additional sorting argument to CitationBibliography().

bib = CitationBibliography("some_references.bib", sorting = :nyt)

The currently supported sorting orders and their behaviours are managed by Bibliography.sorting_rules.

How to cite references in your documentation

You can now use citations anywhere in your docs!

Here I cite Mayer2012, a useful article, via [Mayer2012](@cite): H. C. Mayer , R. Krechetnikov (2012).

I also cite useful_proof, a book you might not want to read, with [useful_proof](@cite) Alfred North Whitehead , Bertrand Russell (1925).

Some TeX diacritical marks are supported, e.g. Karl Heun (1900).

Clicking on the citations takes you to the bibliography.

Example

DocumenterCitations.Example — Type

Example

Foo.

H. C. Mayer , R. Krechetnikov (2012)